Make sure we don't always pull in modules from installed mkspecs
When building Qt, qmake should not pull in mkspecs from an installed
version, as this will lead to incorrect paths for the current build.

So, here we ensure only to pull in modules from there we detected the
qconfig.pri, and ensure we still handle the QMAKEPATH case used by
some.