Remove Qt3Support code from algorithms autotest
qHeapSort and qBubbleSort were Qt3 functions that were replaced by
qSort.  During modularization, the Qt3Support header containing these
routines (q3tl.h) was moved into the algorithms autotest, presumably
because that was the only thing that still included the header.
However, as these routines are not part of Qt5, they don't need to be
tested anymore.  This commit deletes the q3tl.h header and the test
functions that tested it.

Change-Id: I52eed7b4b4db3bd671dc6dbd752642bc9e783c30
Task-number: QTBUG-19325
Reviewed-by: Rohan McGovern