Fixed ordering problem when configuring with `-make <part>'
Prior to this change, running configure with the `-make' option would
affect the order in which parts of Qt are built.  This is unintuitive
and would easily cause build failures.  For example, configuring with
`./configure -make demos' would attempt to build demos before building
libs, which, of course, would fail.

Refactor the code so that the result is the same regardless of the order
of `-make' options.

Change-Id: Idfa61834a0f01d0628a9a1ae27ece94ae3647e6d
Reviewed-on: http://codereview.qt.nokia.com/128
Reviewed-by: Lincoln Ramsay <lincoln.ramsay@nokia.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>