Make it possible to update a related table after an external update
When a table that is related to in a QSqlRelationalTableModel gets
updated in some way (e.g. a new row, or the data is changed) then the
related model could not be updated without recreating the
QSqlRelationalTableModel.

Now, to get around this, select() can be called on the related model to
get it to be updated.

Task-number: QTBUG-7885
Reviewed-by: Charles Yin
Reviewed-by: Michael Goddard

Change-Id: Ic589e840234f3a809bcb112a807a87afe0bc25ca
(cherry picked from commit 2c60a4f67f9fb02f3b711fe749b2f293a07b4e02)
Reviewed-on: http://codereview.qt.nokia.com/2224
Reviewed-by: Charles Yin <charles.yin@nokia.com>