Fixed opengldesktop configure test for mac.
This configure test has never worked for mac.

In a non-qpa configuration, this didn't matter, as the configure script
would skip this configure test on mac.  The configure script could also
be modified to skip the configure test in the qpa case, but it seems
cleaner to make the test actually work.

Change-Id: I0402c490fc4c2401f6363ba2215b261f74ee5d75
Reviewed-on: http://codereview.qt.nokia.com/1127
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Jonas Rabbe <jonas.rabbe@nokia.com>