Support non-desktop platforms better with QDragManager.
By default QDragManager brings in a number of "desktopisms" (e.g. that
a drag is always controlled by mouse events, can be cancelled with some
key press, etc.), these assumptions are not necessarily suitable for
all QPA platforms. Therefore there has to be a way to disable this
built-in event handling, and let the QPlaformDrag decide how the drag
is controlled and when it ends.

Change-Id: Ie9fa2df7ba1e30dd1c6e2cff406490e36be73c4f
Reviewed-on: http://codereview.qt.nokia.com/2085
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Samuel Rødal <samuel.rodal@nokia.com>