test: fixed XPASS from tst_qstatictext on qpa
These tests were written to compare two different methods of drawing
text, to ensure the result was the same.  However, if _both_ methods
turned out to be complete no-ops, the test would unexpectedly pass, as
it never verified that anything was painted.  Add additional
verification steps and QEXPECT_FAIL them on qpa for now.

Task-number: QTBUG-20977

Change-Id: I1ce1b7c2bcd683fc9eeaeb8d1483268e56f5cf08
Reviewed-on: http://codereview.qt.nokia.com/3117
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>