tests: disabled failing qsharedmemory, qsystemsemaphore tests
The `lackey' helper app was disabled by
ccd5228a9380fbfa308a57365e4075f90010da00, but the tests which depend on
it weren't disabled - so, of course, they fail.

None of these tests will work until the circular dependency is fixed
(the tests are in qtbase and they depend on qtscript which depends on
qtbase).

Change-Id: I8e41b4182f5ea57a71ec810a30a21ba1c7217db3
Reviewed-on: http://codereview.qt.nokia.com/3122
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>