Fix QPixmap::grabWindow() on Mac OS X Lion
The old approach was not the recommended one and although it worked on
Mac OS X 10.6 it did not work on 10.7.  The new approach works correctly
on 10.6 and 10.7, so the fix is applied for both versions.

Task-number: QTBUG-19824
Merge-request: 1332
Reviewed-by: sroedal

(cherry picked from commit aaf94776ce1034ec2b2402b9ca0e5cf7c7848b12)

Change-Id: I567d7e2ed2e4b3256f3110f4110ab1342087bc43
Reviewed-on: http://codereview.qt-project.org/4583
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>