Enable multiple instances of QAbstractTestLogger.
Previously QAbstractTestLogger used a global variable for the file
pointer to which it was writing test output.  This effectively meant
that only one instance of this or its derived classes could exist at any
time.

This commit moves the file pointer inside the class, so that multiple
loggers can exist at the same time.  This means that the outputString()
method can no longer be static, which in turn means that several
functions used by QPlainTestLogger need to move from the QTest namespace
into the class, and also that QTestBasicStreamer must hold a non-const
pointer to its associated logger instead of a const pointer.

Task-number: QTBUG-20615
Change-Id: If941f1f9399cf20fb93e3e87f3390bceeca1cbfc
Reviewed-on: http://codereview.qt.nokia.com/3576
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>