test: removed obsolete tst_maketestselftest::tests_auto_pro
This test attempted to ensure that tests/auto/auto.pro would only refer
to other .pro files, attempting to guarantee that the set of all tests
could be cleanly broken up.  The purpose of this was to enable CI
optimizations (e.g. running the autotests for different modules on
different machines in parallel).

This test is invalidated by pending commits which rearrange the
autotests, and we never made use of this property for optimization
anyway, so drop this part of the test.

Change-Id: I23e712fb8ec1dbe7ac65fe66015e1f060f3dcd41
Reviewed-on: http://codereview.qt.nokia.com/3681
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Jo Asplin <jo.asplin@nokia.com>