Fix inclusion of .qmake.cache for modules with config tests.
Qt modules with config tests generate a module level .qmake.cache with
the results of the config tests.  The existance of this file prevents
qmake from walking up the directory tree and finding the Qt global
cache file.  This results in the system build not working as expected.
For example running 'make' in the module would only build the contents
of src and skip tests and examples.

Fixed by adding a include statement to the end of the config test
generated .qmake.cache.

Change-Id: I68a5f2a96f4ee02076b6457ae085f45d894cd4ed
Reviewed-on: http://codereview.qt.nokia.com/3830
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Michael Goddard <michael.goddard@nokia.com>