test: fixed `assert' selftest for release-only mac builds
This test intentionally crashes itself by using Q_ASSERT.

Since Q_ASSERT is a no-op in release builds, the .pro file was
attempting to always force debug mode for this project.

However, on platforms where debug vs release affect linkage (e.g. mac
and windows), this is invalid.  On these platforms, debug mode can't be
enabled unless debug versions of the Qt libraries are available, which
they are not if Qt is configured with `-no-debug-and-release -release'.

Use a different method to ensure that Q_ASSERT really asserts, for both
debug and release builds.

Change-Id: I13eea2c72c77a0a981850dbcaa77f65f147c8490
Reviewed-on: http://codereview.qt.nokia.com/4015
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Jason McDonald <jason.mcdonald@nokia.com>