Try to ensure that QT_ARCH is always set when configuring v8
This is a workaround for QTBUG-21224.

For v8 we want to use QT_ARCH as the primary variable for
detecting the target platform. Unfortunately, QT_ARCH isn't
set when v8.pro is parsed using fromfile() from within
another .pro file (namely src/src.pro).

qt_config seems to be the bare minimum that's needed to get
qconfig.pri (which sets QT_ARCH and friends) loaded.

Change-Id: Ideb713724b98dd100560eaf6d7be39df9a22a71e
Reviewed-on: http://codereview.qt.nokia.com/4094
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>
Reviewed-by: Marius Storm-Olsen <marius.storm-olsen@nokia.com>